Make transactions for immediately consistent events explicit.
