+++ src/Surfnet/StepupMiddleware/CommandHandlingBundle/Resources/config/command_handlers.yml
@@ -3,7 +3,9 @@ services:
         class: Surfnet\StepupMiddleware\CommandHandlingBundle\Identity\CommandHandler\IdentityCommandHandler
         arguments:
             - @surfnet_stepup.repository.identity
-            - @surfnet_stepup_middleware_command_handling.event_bus.transaction_aware_event_flusher
+            - @surfnet_stepup_middleware_command_handling.event_bus.buffered
+            - @doctrine.dbal.middleware_connection
+            - @doctrine.dbal.gateway_connection
         tags: [{ name: command_bus.command_handler }]
 
     surfnet_stepup_middleware_command_handling.command_handler.configuration:
@@ -13,5 +15,7 @@ services:
             - @surfnet_stepup_middleware_gateway.service.gateway_configuration
             - @surfnet_stepup_middleware_api.projector.raa
             - @surfnet_stepup_middleware_api.projector.sraa
+            - @doctrine.dbal.middleware_connection
+            - @doctrine.dbal.gateway_connection
         tags: [{ name: command_bus.command_handler }]
